Update frontend build dependencies & build chain
Laravel mix makes things hard to upgrade (due to dependencies of
dependencies, etc.), and we don't need most of its features. Switch for
plain Webpack which is probably more usable since the module was set up.

Configure and run Prettier to enforce a coding style, don't worry about
other lints yet because the project is so small.